# backend/docs/swagger.json
{
  "schemes": ["http"],
  "swagger": "2.0",
  "info": {
    "description": "AI Agent Demo — Go + Gin 后端，提供 13 个技术演示场景的 API 支持",
    "title": "Interview Demo API",
    "contact": {
      "name": "Developer",
      "email": "dev@example.com"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "paths": {
...
    "/knowledge-base": {
      "get": {
        "description": "返回所有知识库列表（含文档数、分块数）",
        "produces": ["application/json"],
        "tags": ["知识库"],
        "summary": "列出知识库",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "post": {
        "description": "创建新的知识库，同时创建对应的向量集合",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["知识库"],
        "summary": "创建知识库",
        "parameters": [
          {
            "description": "创建请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "description": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/knowledge-base/search": {
      "post": {
        "description": "在知识库中搜索相关文档（语义搜索）",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["知识库"],
        "summary": "搜索知识库",
        "parameters": [
          {
            "description": "搜索请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "knowledgeBaseId": {
                  "type": "string"
                },
                "query": {
                  "type": "string"
                },
                "topK": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/knowledge-base/{id}": {
      "get": {
        "description": "获取指定知识库的详细信息，包含文档列表",
        "produces": ["application/json"],
        "tags": ["知识库"],
        "summary": "获取知识库详情",
        "parameters": [
          {
            "type": "string",
            "description": "知识库 ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "delete": {
        "description": "删除指定知识库及其向量集合",
        "produces": ["application/json"],
        "tags": ["知识库"],
        "summary": "删除知识库",
        "parameters": [
          {
            "type": "string",
            "description": "知识库 ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/knowledge-base/{id}/document": {
      "get": {
        "description": "获取指定知识库的文档列表",
        "produces": ["application/json"],
        "tags": ["知识库"],
        "summary": "获取文档列表",
        "parameters": [
          {
            "type": "string",
            "description": "知识库 ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "post": {
        "description": "向知识库中添加文档，自动进行分块和向量化存储",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["知识库"],
        "summary": "添加文档",
        "parameters": [
          {
            "type": "string",
            "description": "知识库 ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "文档内容",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string"
                },
                "mimeType": {
                  "type": "string"
                },
                "source": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/knowledge-base/{id}/document/{docId}": {
      "delete": {
        "description": "从知识库中删除指定文档",
        "produces": ["application/json"],
        "tags": ["知识库"],
        "summary": "删除文档",
        "parameters": [
          {
            "type": "string",
            "description": "知识库 ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "文档 ID",
            "name": "docId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/knowledge-base/{id}/documents/batch": {
      "post": {
        "description": "向知识库中批量添加多篇文档，每篇自动分块和向量化",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["知识库"],
        "summary": "批量添加文档",
        "parameters": [
          {
            "type": "string",
            "description": "知识库 ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "批量文档",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "documents": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "content": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
...
    "/payments/create": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "创建支付订单，支持幂等 Key 防止重复创建",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["支付"],
        "summary": "创建支付订单",
        "parameters": [
          {
            "description": "创建支付请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/payment.Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/payments/idempotency-test": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "演示幂等 Key 机制：相同 Key 的重复请求返回缓存结果，不产生重复扣款",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["支付"],
        "summary": "幂等性测试",
        "parameters": [
          {
            "description": "幂等请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "amount": {
                  "type": "integer"
                },
                "key": {
                  "type": "string"
                },
                "orderNo": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/payments/order/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "根据 ID 查询支付订单详情",
        "produces": ["application/json"],
        "tags": ["支付"],
        "summary": "查询订单",
        "parameters": [
          {
            "type": "string",
            "description": "订单 ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/payments/orders": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "返回所有支付订单列表（按创建时间倒序）",
        "produces": ["application/json"],
        "tags": ["支付"],
        "summary": "订单列表",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/payments/process/{id}": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "执行支付扣款（模拟延迟），随机成功/失败",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["支付"],
        "summary": "处理支付扣款",
        "parameters": [
          {
            "type": "string",
            "description": "支付订单 ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/payments/retry-demo": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "演示指数退避重试 + 随机抖动策略",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["支付"],
        "summary": "重试演示",
        "parameters": [
          {
            "description": "重试请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/payment.DemoRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/payments/security-check": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "支付回调安全校验演示（IP 白名单、金额二次校验、签名校验）",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["支付"],
        "summary": "安全校验",
        "parameters": [
          {
            "description": "安全校验请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/payment.CheckRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/payments/transition/{id}": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "按有限状态机规则转换订单（PENDING→PROCESSING→SUCCESS/FAIL→REFUNDING→REFUNDED）",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["支付"],
        "summary": "订单状态转换",
        "parameters": [
          {
            "type": "string",
            "description": "订单 ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "目标状态",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "payment.CheckRequest": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "integer"
        },
        "ip": {
          "type": "string"
        },
        "orderNo": {
          "type": "string"
        },
        "sign": {
          "type": "string"
        }
      }
    },
    "payment.DemoRequest": {
      "type": "object",
      "properties": {
        "maxRetries": {
          "type": "integer"
        }
      }
    },
    "payment.Request": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "integer"
        },
        "channel": {
          "type": "string"
        },
        "idempotencyKey": {
          "type": "string"
        },
        "orderNo": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "输入格式: \"Bearer {token}\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}